Move chain extension to runtime-adjacent file
